# frontend-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDefaultsResponse": {
        "properties": {
          "maxTokens": {
            "title": "Maxtokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "model",
          "provider",
          "maxTokens",
          "temperature"
        ],
        "title": "AgentDefaultsResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "A stable, machine-readable description of an API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The common wire format for API errors.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PresetListResponse": {
        "properties": {
          "presets": {
            "items": {
              "$ref": "#/components/schemas/ProviderPresetResponse"
            },
            "title": "Presets",
            "type": "array"
          }
        },
        "required": [
          "presets"
        ],
        "title": "PresetListResponse",
        "type": "object"
      },
      "PresetSyncResponse": {
        "properties": {
          "downloaded": {
            "title": "Downloaded",
            "type": "integer"
          }
        },
        "required": [
          "downloaded"
        ],
        "title": "PresetSyncResponse",
        "type": "object"
      },
...
      "ProviderPresetResponse": {
        "properties": {
          "baseUrl": {
            "title": "Baseurl",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "defaultModel": {
            "title": "Defaultmodel",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "websiteUrl": {
            "title": "Websiteurl",
            "type": "string"
          }
        },
        "required": [
          "name",
          "provider",
          "category",
          "baseUrl",
          "defaultModel",
          "websiteUrl",
          "description"
        ],
        "title": "ProviderPresetResponse",
        "type": "object"
      },
      "ProviderSettingsCreate": {
        "additionalProperties": false,
        "properties": {
          "apiBase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apibase"
          },
          "apiKey": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          },
          "defaultModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultmodel"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "name",
          "provider"
        ],
        "title": "ProviderSettingsCreate",
        "type": "object"
      },
      "ProviderSettingsResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "apiBase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apibase"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "defaultModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultmodel"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "enabled",
          "configured",
          "apiBase",
          "defaultModel",
          "active"
        ],
        "title": "ProviderSettingsResponse",
        "type": "object"
      },
      "ProviderSettingsUpdate": {
        "additionalProperties": false,
        "properties": {
          "apiBase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apibase"
          },
          "apiKey": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          },
          "defaultModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultmodel"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "title": "ProviderSettingsUpdate",
        "type": "object"
      },
...
      "SettingsDefaultsUpdate": {
        "additionalProperties": false,
        "properties": {
          "activeProviderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activeproviderid"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "title": "SettingsDefaultsUpdate",
        "type": "object"
      },
      "SettingsResponse": {
        "properties": {
          "defaults": {
            "$ref": "#/components/schemas/AgentDefaultsResponse"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderSettingsResponse"
            },
            "title": "Providers",
            "type": "array"
          }
        },
        "required": [
          "providers",
          "defaults"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
      "SettingsValidationResponse": {
        "properties": {
          "availableProviders": {
            "items": {
              "type": "string"
            },
            "title": "Availableproviders",
            "type": "array"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "availableProviders",
          "errors"
        ],
        "title": "SettingsValidationResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ControlLeaseToken": {
        "in": "header",
        "name": "X-Control-Lease-Token",
        "type": "apiKey"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "manyselves_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "get_api_v1_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Get Settings"
      },
      "patch": {
        "operationId": "patch_api_v1_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsDefaultsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Update Defaults"
      }
    },
    "/api/v1/settings/presets": {
      "get": {
        "operationId": "get_api_v1_settings_presets",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Provider Presets"
      }
    },
    "/api/v1/settings/presets/sync": {
      "post": {
        "operationId": "post_api_v1_settings_presets_sync",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Synchronize Provider Presets"
      }
    },
    "/api/v1/settings/providers": {
      "post": {
        "operationId": "post_api_v1_settings_providers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSettingsCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Create Provider"
      }
    },
    "/api/v1/settings/providers/{provider_id}": {
      "delete": {
        "operationId": "delete_api_v1_settings_providers_provider_id",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Remove Provider"
      },
      "patch": {
        "operationId": "patch_api_v1_settings_providers_provider_id",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Update Provider"
      }
    },
    "/api/v1/settings/validate": {
      "post": {
        "operationId": "post_api_v1_settings_validate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Validate Settings"
      }
    }
  }
}